+++ app/src/main/kotlin/br/com/messore/tech/turtleracing/core/runCatchingWithDispatcher.kt
@@ -7,7 +7,7 @@ import timber.log.Timber
 suspend fun <T, R> T.runCatchingWithDispatcher(
     dispatcher: CoroutineDispatcher,
     execute: suspend () -> R,
-    onSuccess: (R) -> Unit,
+    onSuccess: (R) -> Unit = {},
     onFailure: (Throwable) -> Unit = {}
 ) {
     runCatching {
