--- app/src/main/kotlin/br/com/messore/tech/turtleracing/presentation/login/LoginUiAction.kt
@@ -1,6 +1,0 @@
-package br.com.messore.tech.turtleracing.presentation.login
-
-sealed class LoginUiAction {
-    object Logged : LoginUiAction()
-    data class ErrorOnLogin(val message: String) : LoginUiAction()
-}
