--- app/src/main/kotlin/br/com/messore/tech/turtleracing/presentation/login/LoginUiState.kt
@@ -1,5 +1,0 @@
-package br.com.messore.tech.turtleracing.presentation.login
-
-data class LoginUiState(
-    val isLoading: Boolean = false
-)
