+++ app/src/main/kotlin/br/com/messore/tech/turtleracing/presentation/turtle/TurtleUiAction.kt
@@ -1,0 +1,6 @@
+package br.com.messore.tech.turtleracing.presentation.turtle
+
+sealed class TurtleUiAction {
+    object Logged : TurtleUiAction()
+    data class ErrorOnLogin(val message: String) : TurtleUiAction()
+}
