+++ app/src/main/kotlin/br/com/messore/tech/turtleracing/presentation/turtle/TurtleUiState.kt
@@ -1,0 +1,5 @@
+package br.com.messore.tech.turtleracing.presentation.turtle
+
+data class TurtleUiState(
+    val isLoading: Boolean = false
+)
